Checking for valid codes
Created code for assignment 3. A function will scan a file and determine whether each line of product codes are valid based off criteria provided (longer than 10 characters, #4-7 are digits, and #10 is an uppercase letter). Another function will take each valid code and check whether they are restricted. This will check the 10th character to be 'R', and than check if the area code is above 2000